backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pydantic==2.*
pydantic[email]
beautifulsoup4
lxml
mysql-connector-python
openai
bcrypt
pyjwt